requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
langchain-community
flask
pypdf
python-dotenv
-e .